Commited to initialising SizeClass::heterotrophs_ to the possible maximum based on the initial parameters. This has the potential to require the least amount of CPU time, but requires the most memory. This approach has led to a rename of OutputData::livingCount_ and the introduction of another variable track the reciprocal.

+++ src/Model/Heterotrophs.cpp
@@ -22,7 +22,7 @@ Heterotrophs::Heterotrophs(Nutrient& nutrient, Parameters& params, const double&
   alive_.reserve(maxPopulation);
 }
 
-std::size_t Heterotrophs::getPopulationSize() {
+std::uint64_t Heterotrophs::getPopulationSize() {
   return alive_.size();
 }
